完成部分接口，完善代码结构

+++ src/main/java/com/sc/qisi_system/module/admin/dto/MenuQueryDTO.java
@@ -29,5 +29,5 @@ public class MenuQueryDTO {
      * 业务身份ID / 角色ID
      * 对应 sys_identity.id
      */
-    private Long identityId;
+    private Integer identityId;
 }
